Use database replicas for read requests
Fixes #2634

The connection switching is managed using the standard Resolver, which
chooses the connection based on the HTTP verb. In the small number of cases
where we update the database during GET requests, we override the
default choice of connections.

Additionally, after any write in a particular user session, the resolver
will continue to use the same (write) connection for two seconds for
subsequent reads. This gives time for the changes to be replicated before
switching back.