Use squasher to combine old migrations into one
There's no need to keep the exact history of all the migrations,
that's available in the repository history. They contain code that
needs to be maintained, and some of the code is no longer required.

This combines all the migrations up until 2018 into one "starter"
migration. If you need to migrate a database older than this, then
you will need to checkout an older version of the codebase.